button function adjusted
