+++ src/components/Recipe.jsx
@@ -3,15 +3,13 @@ import {useContentful} from 'react-contentful'
 import {client} from '../client.js';
 import './recipe.css'
 import ReactMarkdown from 'react-markdown';
+import RecipeCard from './RecipeCard.jsx';
 
 
 export default function Recipe() {
 
 const [recipesData, setRecipesData] = useState([]);
-const [showPrep, setShowPrep] = useState(false);
-const handleShowPrep = () => !showPrep ? setShowPrep(true) : setShowPrep(false);
-const [showIngredients, setShowIngredients] = useState(false);
-const handleShowIngredients = () => !showIngredients ? setShowIngredients(true): setShowIngredients(false);
+
 
 useEffect(() => {
 
@@ -21,45 +19,29 @@ useEffect(() => {
     );
      },[]);
 
+console.log(recipesData);
 
 return(
 <>
     
-   <div className='RecipesOverview'>
+   <div>
    
        {recipesData.map((recipes) => (
-    <div className='Recipe'>
-        
-        <div className='title'> {recipes.fields.title} </div>
-        <ReactMarkdown className='description'>
-        {recipes.fields.description}
-        </ReactMarkdown>
-        
-        <ReactMarkdown className='rating'>
-        {recipes.fields.rating}
-        </ReactMarkdown>
-       <div className='image'> <img src={recipes.fields.picture.fields.file.url} alt="Food"/> </div>
-       <div>
-            <button className='btnShow' onClick={handleShowIngredients}>{!showIngredients ? "Show Ingredients" : "Hide List"}</button>
-            {showIngredients ? 
-            <ReactMarkdown className='ingredients'>
-            {recipes.fields.ingredients}
-            </ReactMarkdown> : null }
-        </div>
-       <div> 
-            <button className='btnShow' onClick={handleShowPrep}>{!showPrep ? "Show Preperation" : "Close Liste"}</button>
-            {showPrep ? 
-            <ReactMarkdown className='prep'>
+           <RecipeCard
+           title={recipes.fields.title}
+           description={recipes.fields.description}
+            rating={recipes.fields.rating}
+            image={recipes.fields.picture.fields.file.url}
+            ingredients={recipes.fields.ingredients}
+            preperation=
             {recipes.fields.prep}
-            </ReactMarkdown> : null }
-        </div>
-        
-    </div>   
-       ))}
-    </div>
-</>
+           />
+           ))}
+           </div>
+         </>
+       );
+   
 
-)
 }
